perf(fe): Fixed some code structure based on the comments
