+++ apps/web/src/app/constants/expertise.constants.ts
@@ -1,0 +1,21 @@
+/**
+ * Standard expertise options for mentor registration
+ */
+export const EXPERTISE_OPTIONS = [
+  'Software Engineering',
+  'Frontend Development',
+  'Backend Development', 
+  'Full-Stack Development',
+  'DevOps & Infrastructure',
+  'Cloud Architecture',
+  'Data Science & Analytics',
+  'Machine Learning & AI',
+  'Product Management',
+  'UI/UX Design',
+  'Cybersecurity',
+  'Mobile Development',
+  'Database Design',
+  'System Architecture',
+  'Quality Assurance',
+  'Project Management'
+];
