+++ apps/web/src/app/constants/timezone-mapping.constants.ts
@@ -1,0 +1,119 @@
+/**
+ * Country to default timezone mapping
+ * Based on capital/major cities
+ */
+export const COUNTRY_TIMEZONES: Record<string, string> = {
+  // Asia Pacific
+  PH: 'Asia/Manila',
+  Philippines: 'Asia/Manila',
+  SG: 'Asia/Singapore',
+  Singapore: 'Asia/Singapore',
+  MY: 'Asia/Kuala_Lumpur',
+  Malaysia: 'Asia/Kuala_Lumpur',
+  TH: 'Asia/Bangkok',
+  Thailand: 'Asia/Bangkok',
+  VN: 'Asia/Ho_Chi_Minh',
+  Vietnam: 'Asia/Ho_Chi_Minh',
+  ID: 'Asia/Jakarta',
+  Indonesia: 'Asia/Jakarta',
+  JP: 'Asia/Tokyo',
+  Japan: 'Asia/Tokyo',
+  KR: 'Asia/Seoul',
+  'South Korea': 'Asia/Seoul',
+  CN: 'Asia/Shanghai',
+  China: 'Asia/Shanghai',
+  IN: 'Asia/Kolkata',
+  India: 'Asia/Kolkata',
+  PK: 'Asia/Karachi',
+  Pakistan: 'Asia/Karachi',
+  BD: 'Asia/Dhaka',
+  Bangladesh: 'Asia/Dhaka',
+  AU: 'Australia/Sydney',
+  Australia: 'Australia/Sydney',
+  NZ: 'Pacific/Auckland',
+  'New Zealand': 'Pacific/Auckland',
+  
+  // Americas
+  US: 'America/New_York',
+  'United States': 'America/New_York',
+  CA: 'America/Toronto',
+  Canada: 'America/Toronto',
+  MX: 'America/Mexico_City',
+  Mexico: 'America/Mexico_City',
+  BR: 'America/Sao_Paulo',
+  Brazil: 'America/Sao_Paulo',
+  AR: 'America/Argentina/Buenos_Aires',
+  Argentina: 'America/Argentina/Buenos_Aires',
+  CL: 'America/Santiago',
+  Chile: 'America/Santiago',
+  CO: 'America/Bogota',
+  Colombia: 'America/Bogota',
+  
+  // Europe
+  GB: 'Europe/London',
+  'United Kingdom': 'Europe/London',
+  UK: 'Europe/London',
+  DE: 'Europe/Berlin',
+  Germany: 'Europe/Berlin',
+  FR: 'Europe/Paris',
+  France: 'Europe/Paris',
+  IT: 'Europe/Rome',
+  Italy: 'Europe/Rome',
+  ES: 'Europe/Madrid',
+  Spain: 'Europe/Madrid',
+  NL: 'Europe/Amsterdam',
+  Netherlands: 'Europe/Amsterdam',
+  BE: 'Europe/Brussels',
+  Belgium: 'Europe/Brussels',
+  SE: 'Europe/Stockholm',
+  Sweden: 'Europe/Stockholm',
+  NO: 'Europe/Oslo',
+  Norway: 'Europe/Oslo',
+  DK: 'Europe/Copenhagen',
+  Denmark: 'Europe/Copenhagen',
+  FI: 'Europe/Helsinki',
+  Finland: 'Europe/Helsinki',
+  PL: 'Europe/Warsaw',
+  Poland: 'Europe/Warsaw',
+  RU: 'Europe/Moscow',
+  Russia: 'Europe/Moscow',
+  TR: 'Europe/Istanbul',
+  Turkey: 'Europe/Istanbul',
+  
+  // Middle East & Africa
+  AE: 'Asia/Dubai',
+  'United Arab Emirates': 'Asia/Dubai',
+  SA: 'Asia/Riyadh',
+  'Saudi Arabia': 'Asia/Riyadh',
+  IL: 'Asia/Jerusalem',
+  Israel: 'Asia/Jerusalem',
+  ZA: 'Africa/Johannesburg',
+  'South Africa': 'Africa/Johannesburg',
+  EG: 'Africa/Cairo',
+  Egypt: 'Africa/Cairo',
+  KE: 'Africa/Nairobi',
+  Kenya: 'Africa/Nairobi',
+  NG: 'Africa/Lagos',
+  Nigeria: 'Africa/Lagos',
+};
+
+/**
+ * Common timezone options grouped by region
+ */
+export const TIMEZONE_OPTIONS = [
+  { label: 'Asia/Manila (GMT+8)', value: 'Asia/Manila', offset: '+08:00' },
+  { label: 'Asia/Singapore (GMT+8)', value: 'Asia/Singapore', offset: '+08:00' },
+  { label: 'Asia/Tokyo (GMT+9)', value: 'Asia/Tokyo', offset: '+09:00' },
+  { label: 'Asia/Shanghai (GMT+8)', value: 'Asia/Shanghai', offset: '+08:00' },
+  { label: 'Asia/Dubai (GMT+4)', value: 'Asia/Dubai', offset: '+04:00' },
+  { label: 'Asia/Kolkata (GMT+5:30)', value: 'Asia/Kolkata', offset: '+05:30' },
+  { label: 'Europe/London (GMT+0)', value: 'Europe/London', offset: '+00:00' },
+  { label: 'Europe/Paris (GMT+1)', value: 'Europe/Paris', offset: '+01:00' },
+  { label: 'Europe/Berlin (GMT+1)', value: 'Europe/Berlin', offset: '+01:00' },
+  { label: 'America/New_York (GMT-5)', value: 'America/New_York', offset: '-05:00' },
+  { label: 'America/Chicago (GMT-6)', value: 'America/Chicago', offset: '-06:00' },
+  { label: 'America/Los_Angeles (GMT-8)', value: 'America/Los_Angeles', offset: '-08:00' },
+  { label: 'America/Toronto (GMT-5)', value: 'America/Toronto', offset: '-05:00' },
+  { label: 'Australia/Sydney (GMT+10)', value: 'Australia/Sydney', offset: '+10:00' },
+  { label: 'Pacific/Auckland (GMT+12)', value: 'Pacific/Auckland', offset: '+12:00' },
+];
