+++ apps/web/src/app/interfaces/alert-config.interface.ts
@@ -1,0 +1,8 @@
+import { AlertType } from '../types/alert.types';
+
+export interface AlertConfig {
+  type: AlertType;
+  title: string;
+  message: string;
+  dismissible?: boolean;
+}
