+++ apps/web/src/app/interfaces/country.interface.ts
@@ -1,0 +1,8 @@
+export interface Country {
+  value: string;
+  label: string;
+  phoneCode: string;
+  flag: string;
+  phoneFormat: string;
+  phoneRegex: RegExp;
+}
