+++ apps/web/src/app/interfaces/language.interface.ts
@@ -1,0 +1,4 @@
+export interface Language {
+  value: string;
+  label: string;
+}
