+++ apps/web/src/app/interfaces/step.interface.ts
@@ -1,0 +1,5 @@
+export interface Step {
+  title: string;
+  description: string;
+  fields: string[];
+}
