+++ apps/web/src/app/interfaces/timezone.interface.ts
@@ -1,0 +1,5 @@
+export interface Timezone {
+  value: string;
+  label: string;
+  countries: string[];
+}
