+++ apps/web/src/app/types/alert.types.ts
@@ -1,0 +1,1 @@
+export type AlertType = 'success' | 'error' | 'warning' | 'info';
