--- apps/web/src/app/types/api.types.ts
@@ -1,2 +1,0 @@
-export type { RegisterMenteeRequest, RegisterMentorRequest, ApiResponse } from '@gurokonekt/models';
-export type { AuthResponse } from './auth-api-response.types';
