--- apps/web/src/app/types/auth-api-response.types.ts
@@ -1,11 +1,0 @@
-export interface AuthResponse {
-  user: {
-    id: string;
-    email: string;
-    fullName: string;
-    role: 'mentee' | 'mentor';
-  };
-  accessToken: string;
-  token?: string; // For backward compatibility
-  message: string;
-}
