--- apps/web/src/app/types/register-mentee-request.types.ts
@@ -1,13 +1,0 @@
-export interface RegisterMenteeRequest {
-  firstName: string;
-  lastName: string;
-  middleName?: string;
-  suffix?: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  phoneNumber: string;
-  country: string;
-  timezone: string;
-  language: string;
-}
