+++ lib/models/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/models';
+export * from './lib/interfaces/auth/auth-response.interface';
 export * from './lib/interfaces/auth/signup.model';
 export * from './lib/interfaces/auth/signin.model';
 export * from './lib/interfaces/logs/logs.model';
